Tag Folders: Refactor tag code into tag.js
